requirements.txt
pydash==7.0.1
beautifulsoup4==4.12.2
PyYAML==6.0.1
lxml==4.9.2